+++ package.json
@@ -8,13 +8,11 @@
   ],
   "scripts": {
     "prepare": "bunx lefthook install",
-    "update": "bun upgrade && bun update --latest && bun install",
+    "update": "bun upgrade && bun update --recursive --latest && bun install",
     "rebuild": "rm -rf node_modules && rm -rf packages/frontend/node_modules && rm -rf packages/dev/node_modules && rm -rf packages/frontend/dist && rm -rf packages/frontend/.astro && bun run update",
     "reset": "bun run rebuild && bun run devtools && bun run build",
     "submodule": "git submodule update --recursive --remote",
-    "gsheets2img": "bun --filter gsheets2img main -g",
-    "imgur": "bun --filter gsheets2img main -i",
-    "gsheets2imgur": "bun --filter gsheets2img main -gi",
+    "gsheets2img": "bun --filter gsheets2img main",
     "compress": "bun --filter dev compress",
     "smol-format": "bunx --bun prettier --write \"packages/**/*.{astro,json}\"",
     "format": "bunx --bun prettier --write \"packages/**/*.{astro,html,js,jsx,md,mdx,svelte,ts,tsx,vue,json,css}\"",
