+++ packages/frontend/src/components/Equipment/EquipmentContent.astro
@@ -20,8 +20,14 @@ import EquipmentLinkButtons from "./EquipmentLinkButtons.astro"
   </li>
   <li>
     <p>
-      If the images are not loading, consider using one of the alternative links
-      listed below:
+      If the images are not loading, click the 'Spreadsheet' link below to open
+      a Google Sheets version of this page.
+    </p>
+  </li>
+  <li>
+    <p>
+      For a less comprehensive but simpler gear guide, you can read Suchiguma's
+      Quick Gear Guide linked below.
     </p>
   </li>
 </ul>
