+++ packages/frontend/src/components/Equipment/EquipmentLinkButtons.astro
@@ -33,15 +33,6 @@
   aria-label="Links"
 >
   <a
-    href="https://imgur.com/a/YoxqS7A"
-    target="_blank"
-    rel="noopener noreferrer"
-    role="button"
-    class="equip-link-btn w-full"
-  >
-    Imgur
-  </a>
-  <a
     href="https://docs.google.com/spreadsheets/d/1wWMIzaUKISAXMbOEnmsuuLkO9PesabpdTUWdosvHygM/edit#gid=979150851"
     target="_blank"
     rel="noopener noreferrer"
@@ -49,5 +40,14 @@
     class="equip-link-btn w-full"
   >
     Spreadsheet
+  </a>
+  <a
+    href="https://suchiguma.github.io/miniguides/quick-gear-guide"
+    target="_blank"
+    rel="noopener noreferrer"
+    role="button"
+    class="equip-link-btn w-full"
+  >
+    Quick Gear Guide (by Suchiguma)
   </a>
 </div>
